Adding section-11-project-3-fastapi-authenticate-requests-admin-router

+++ project_3/TodoApp/routers/todos.py
@@ -5,8 +5,12 @@ from ..models import Todos
 from ..database import SessionLocal
 from typing import Annotated
 from sqlalchemy.orm import Session
+from .auth import get_current_user
 
-router = APIRouter()
+router = APIRouter(
+    prefix='/todo',
+    tags=['todos']
+)
 
 
 def get_db():
@@ -18,6 +22,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -27,33 +32,54 @@ class TodoRequest(BaseModel):
     complete: bool
 
 
-@router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+@router.get("", status_code=status.HTTP_200_OK)
+async def read_all(user: user_dependency,
+                   db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
-@router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency,
+@router.get("/{todo_id}", status_code=status.HTTP_200_OK)
+async def read_todo(user: user_dependency,
+                    db: db_dependency,
                     todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
+
     if todo_model is not None:
         return todo_model
+
     raise HTTPException(status_code=404, detail='Todo not found.')
 
 
-@router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency,
+@router.post("", status_code=status.HTTP_201_CREATED)
+async def create_todo(user: user_dependency,
+                      db: db_dependency,
                       todo_request: TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = Todos(**todo_request.model_dump(), owner_id=user.get('id'))
+
     db.add(todo_model)
     db.commit()
 
 
-@router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+@router.put("/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(user: user_dependency, db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
 
@@ -61,15 +87,23 @@ async def update_todo(db: db_dependency,
     todo_model.description = todo_request.description
     todo_model.priority = todo_request.priority
     todo_model.complete = todo_request.complete
+
     db.add(todo_model)
     db.commit()
 
 
-@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency,
+@router.delete("/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency,
+                      db: db_dependency,
                       todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
+
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
